clean up per-unit modifiers (e.g. flagship abilities).  instead of adding modifier code to unit, add a 'unitAbility' and register that in unit-modifiers.data.json the same way we do for faction abilities -- keeps all unit modifier code in the same place.
