+++ src/lib/unit/unit-attrs.data.js
@@ -430,26 +430,7 @@ module.exports = [
         upgradeLevel: 1,
         localeName: "unit.mech.aerie_sentinel",
         triggerNsid: "card.leader.mech.argent:pok/aerie_sentinel",
-        unitModifier: {
-            // Does not count against capacity when with a ship with capacity
-            localeName: "unit.mech.aerie_sentinel",
-            localeDescription: "unit_modifier.desc.aerie_sentinel",
-            owner: "self",
-            priority: "mutate",
-            applyAll: (unitAttrsSet, auxData) => {
-                let hasCapacity = false;
-                for (const unitAttrs of unitAttrsSet.values()) {
-                    if (unitAttrs.raw.capacity) {
-                        hasCapacity = true;
-                        break;
-                    }
-                }
-                if (hasCapacity) {
-                    const mechAttrs = unitAttrsSet.get("mech");
-                    delete mechAttrs.raw.requireCapacity;
-                }
-            },
-        },
+        unitAbility: "unit.mech.aerie_sentinel",
     },
     {
         unit: "mech",
@@ -501,27 +482,7 @@ module.exports = [
         upgradeLevel: 1,
         localeName: "unit.mech.iconoclast",
         triggerNsid: "card.leader.mech.naalu:base/iconoclast",
-        unitModifier: {
-            // "+2 mech COMBAT rolls if opponent has fragment",
-            isCombat: true,
-            localeName: "unit.mech.iconoclast",
-            localeDescription: "unit_modifier.desc.iconoclast",
-            owner: "self",
-            priority: "adjust",
-            triggerNsid: "card.leader.mech.naalu:pok/iconoclast",
-            applyAll: (unitAttrsSet, auxData) => {
-                let opponentHasFragment = false; // TODO XXX
-                if (opponentHasFragment) {
-                    const mechAttrs = unitAttrsSet.get("mech");
-                    if (mechAttrs.raw.spaceCombat) {
-                        mechAttrs.raw.spaceCombat.hit -= 2;
-                    }
-                    if (mechAttrs.raw.groundCombat) {
-                        mechAttrs.raw.groundCombat.hit -= 2;
-                    }
-                }
-            },
-        },
+        unitAbility: "unit.mech.iconoclast",
     },
     {
         unit: "mech",
@@ -543,49 +504,14 @@ module.exports = [
         upgradeLevel: 1,
         localeName: "unit.mech.moll_terminus",
         triggerNsid: "card.leader.mech.mentak:base/moll_terminus",
-        unitModifier: {
-            // Other's ground forces on planet cannot SUSTAIN DAMAGE
-            isCombat: true,
-            localeName: "unit.mech.moll_terminus",
-            localeDescription: "unit_modifier.desc.moll_terminus",
-            owner: "opponent",
-            priority: "mutate",
-            applyEach: (unitAttrs, auxData) => {
-                if (
-                    unitAttrs.raw.ground &&
-                    unitAttrs.raw.sustainDamage &&
-                    auxData.opponent.has("mech")
-                ) {
-                    delete unitAttrs.raw.sustainDamage;
-                }
-            },
-        },
+        unitAbility: "unit.mech.moll_terminus",
     },
     {
         unit: "mech",
         upgradeLevel: 1,
         localeName: "unit.mech.mordred",
         triggerNsid: "card.leader.mech.nekro:base/mordred",
-        unitModifier: {
-            // +2 mech COMBAT rolls if opponent has X/Y token
-            isCombat: true,
-            localeName: "unit.mech.mordred",
-            localeDescription: "unit_modifier.desc.mordred",
-            owner: "self",
-            priority: "adjust",
-            applyAll: (unitAttrsSet, auxData) => {
-                let opponentHasXYToken = false; // TODO XXX
-                if (opponentHasXYToken) {
-                    const mechAttrs = unitAttrsSet.get("mech");
-                    if (mechAttrs.raw.spaceCombat) {
-                        mechAttrs.raw.spaceCombat.hit -= 2;
-                    }
-                    if (mechAttrs.raw.groundCombat) {
-                        mechAttrs.raw.groundCombat.hit -= 2;
-                    }
-                }
-            },
-        },
+        unitAbility: "unit.mech.mordred",
     },
     {
         unit: "mech",
@@ -628,36 +554,7 @@ module.exports = [
         upgradeLevel: 1,
         localeName: "unit.mech.shield_paling",
         triggerNsid: "card.leader.mech.jolnar:base/shield_paling",
-        unitModifier: {
-            // Infantry on planet with mech are not FRAGILE
-            isCombat: true,
-            localeName: "unit.mech.shield_paling",
-            localeDescription: "unit_modifier.desc.shield_paling",
-            owner: "self",
-            priority: "adjust",
-            applyAll: (unitAttrsSet, auxData) => {
-                // Normally mech is paired with Jol-Nar + FRAGILE, but watch out for Franken!
-                let hasFragile = false;
-                for (const unitModifier of auxData.self.unitModifiers) {
-                    if (
-                        unitModifier.raw.localeName ==
-                        "unit_modifier.name.fragile"
-                    ) {
-                        hasFragile = true;
-                        break;
-                    }
-                }
-                // Do not attempt to suppress "fragile" application, just undo it.
-                const infantryAttrs = unitAttrsSet.get("infantry");
-                if (
-                    hasFragile &&
-                    auxData.self.has("mech") &&
-                    infantryAttrs.raw.groundCombat
-                ) {
-                    infantryAttrs.raw.groundCombat -= 1;
-                }
-            },
-        },
+        unitAbility: "unit.mech.shield_paling",
     },
     {
         unit: "mech",
@@ -722,26 +619,7 @@ module.exports = [
         triggerNsid:
             "card.technology.unit_upgrade.mahact:franken.pok/arvicon_rex",
         spaceCombat: { dice: 2, hit: 5 },
-        unitModifier: {
-            // +2 flagship COMBAT against opponent with no token in your fleet pool
-            isCombat: true,
-            localeName: "unit.flagship.arvicon_rex",
-            localeDescription: "unit_modifier.desc.arvicon_rex",
-            owner: "self",
-            priority: "adjust",
-            applyAll: (unitAttrsSet, auxData) => {
-                let opponentTokenInFleetPool = false; // TODO XXX
-                if (opponentTokenInFleetPool) {
-                    const flagshipAttrs = unitAttrsSet.get("flagship");
-                    if (flagshipAttrs.raw.spaceCombat) {
-                        flagshipAttrs.raw.spaceCombat.hit -= 2;
-                    }
-                    if (flagshipAttrs.raw.groundCombat) {
-                        flagshipAttrs.raw.groundCombat.hit -= 2;
-                    }
-                }
-            },
-        },
+        unitAbility: "unit.flagship.arvicon_rex",
     },
     {
         unit: "flagship",
@@ -750,23 +628,7 @@ module.exports = [
         triggerNsid:
             "card.technology.unit_upgrade.norr:franken.base/cmorran_norr",
         spaceCombat: { dice: 2, hit: 6 },
-        unitModifier: {
-            // +1 to all COMBAT rolls for other ships with the C'morran N'orr
-            isCombat: true,
-            localeName: "unit.flagship.cmorran_norr",
-            localeDescription: "unit_modifier.desc.cmorran_norr",
-            owner: "self",
-            priority: "adjust",
-            applyEach: (unitAttrs, auxData) => {
-                if (
-                    unitAttrs.raw.ship &&
-                    unitAttrs.raw.unit !== "flagship" &&
-                    unitAttrs.raw.spaceCombat
-                ) {
-                    unitAttrs.raw.spaceCombat.hit -= 1;
-                }
-            },
-        },
+        unitAbility: "unit.flagship.cmorran_norr",
     },
     {
         unit: "flagship",
@@ -791,19 +653,7 @@ module.exports = [
         triggerNsid:
             "card.technology.unit_upgrade.mentak:franken.base/fourth_moon",
         spaceCombat: { dice: 2, hit: 7 },
-        unitModifier: {
-            // Opponent's ships cannot use SUSTAIN DAMAGE
-            isCombat: true,
-            localeName: "unit.flagship.fourth_moon",
-            localeDescription: "unit_modifier.desc.fourth_moon",
-            owner: "opponent",
-            priority: "adjust",
-            applyEach: (unitAttrs, auxData) => {
-                if (unitAttrs.raw.ship && unitAttrs.raw.sustainDamage) {
-                    unitAttrs.raw.sustainDamage = false;
-                }
-            },
-        },
+        unitAbility: "unit.flagship.fourth_moon",
     },
     {
         unit: "flagship",
@@ -846,26 +696,7 @@ module.exports = [
             "card.technology.unit_upgrade.naalu:franken.base/matriarch",
         spaceCombat: { dice: 2, hit: 9 },
         capacity: 6,
-        unitModifier: {
-            // Fighters may participate in ground combat
-            isCombat: true,
-            localeName: "unit.flagship.matriarch",
-            localeDescription: "unit_modifier.desc.matriarch",
-            owner: "self",
-            priority: "adjust",
-            applyEach: (unitAttrs, auxData) => {
-                if (
-                    unitAttrs.raw.unit === "fighter" &&
-                    !unitAttrs.raw.groundCombat
-                ) {
-                    unitAttrs.raw.groundCombat = {
-                        dice: unitAttrs.raw.spaceCombat.dice,
-                        hit: unitAttrs.raw.spaceCombat.hit,
-                        anyPlanet: true,
-                    };
-                }
-            },
-        },
+        unitAbility: "unit.flagship.matriarch",
     },
     {
         unit: "flagship",
@@ -898,19 +729,7 @@ module.exports = [
         triggerNsid:
             "card.technology.unit_upgrade.argent:franken.pok/quetzecoatl",
         spaceCombat: { dice: 2, hit: 7 },
-        unitModifier: {
-            // Other players cannot use SPACE CANNON against your ships in this system
-            isCombat: true,
-            localeName: "unit.flagship.quetzecoatl",
-            localeDescription: "unit_modifier.desc.quetzecoatl",
-            owner: "opponent",
-            priority: "adjust",
-            applyEach: (unitAttrs, auxData) => {
-                if (unitAttrs.raw.spaceCannon) {
-                    delete unitAttrs.raw.spaceCannon;
-                }
-            },
-        },
+        unitAbility: "unit.flagship.quetzecoatl",
     },
     {
         unit: "flagship",
@@ -919,33 +738,7 @@ module.exports = [
         triggerNsid:
             "card.technology.unit_upgrade.winnu:franken.base/salai_sai_corian",
         spaceCombat: { dice: 1, hit: 7 },
-        unitModifier: {
-            // Rolls number of dice equal to number of opponent's non-fighter ships
-            isCombat: true,
-            localeName: "unit.flagship.salai_sai_corian",
-            localeDescription: "unit_modifier.desc.salai_sai_corian",
-            owner: "self",
-            priority: "adjust",
-            applyAll: (unitAttrsSet, auxData) => {
-                let nonFighterShipCount = 0;
-                // TODO XXX NazRhoka mech on planet vs space (count as ship)
-                for (const unitAttrs of auxData.opponent.unitAttrsSet.values()) {
-                    if (
-                        unitAttrs.raw.ship &&
-                        unitAttrs.raw.unit !== "fighter" &&
-                        auxData.opponent.has(unitAttrs.raw.unit)
-                    ) {
-                        nonFighterShipCount += auxData.opponent.count(
-                            unitAttrs.raw.unit
-                        );
-                    }
-                }
-                const flagshipAttrs = unitAttrsSet.get("flagship");
-                if (flagshipAttrs.raw.spaceCombat) {
-                    flagshipAttrs.raw.spaceCombat.dice = nonFighterShipCount;
-                }
-            },
-        },
+        unitAbility: "unit.flagship.salai_sai_corian",
     },
     {
         unit: "flagship",
@@ -963,26 +756,7 @@ module.exports = [
         triggerNsid:
             "card.technology.unit_upgrade.nekro:franken.base/the_alastor",
         spaceCombat: { dice: 2, hit: 9 },
-        unitModifier: {
-            // Ground forces may participate in space combat
-            isCombat: true,
-            localeName: "unit.flagship.the_alastor",
-            localeDescription: "unit_modifier.desc.the_alastor",
-            owner: "self",
-            priority: "adjust",
-            applyEach: (unitAttrs, auxData) => {
-                if (
-                    unitAttrs.raw.ground &&
-                    unitAttrs.raw.groundCombat &&
-                    !unitAttrs.raw.spaceCombat
-                ) {
-                    unitAttrs.raw.spaceCombat = {
-                        dice: unitAttrs.raw.groundCombat.dice,
-                        hit: unitAttrs.raw.groundCombat.hit,
-                    };
-                }
-            },
-        },
+        unitAbility: "unit.flagship.the_alastor",
     },
     {
         unit: "flagship",
@@ -1016,23 +790,7 @@ module.exports = [
             "card.technology.unit_upgrade.naazrokha:franken.pok/visz_el_vir",
         spaceCombat: { dice: 2, hit: 9 },
         capacity: 4,
-        unitModifier: {
-            // Your mechs in this system roll 1 additional die during combat
-            isCombat: true,
-            localeName: "unit.flagship.visz_el_vir",
-            localeDescription: "unit_modifier.desc.visz_el_vir",
-            owner: "self",
-            priority: "adjust",
-            applyEach: (unitAttrs, auxData) => {
-                if (
-                    unitAttrs.raw.unit === "mech" &&
-                    unitAttrs.raw.groundCombat
-                ) {
-                    unitAttrs.raw.groundCombat.extraDice =
-                        (unitAttrs.raw.groundCombat.extraDice || 0) + 1;
-                }
-            },
-        },
+        unitAbility: "unit.flagship.visz_el_vir",
     },
     {
         unit: "flagship",
