+++ src/lib/unit/unit-modifier.schema.js
@@ -9,6 +9,7 @@ const UNIT_MODIFIER_SCHEMA_JSON = {
         triggerNsid: { type: "string" }, // find in player area to apply this unit upgrade
         triggerNsids: { type: "array", items: { type: "string" } },
         triggerFactionAbility: { type: "string" },
+        triggerUnitAbility: { type: "string" },
         owner: { type: "string" }, // self, opponent, or any
         priority: { type: "string" }, // mutate, adjust, or choose
         toggleActive: { type: "boolean" },
